+++ Project/src/main/java/com/lukecreator/BonziBot/Data/UserAccountSort.java
@@ -19,7 +19,7 @@ public class UserAccountSort implements Comparator<Map.Entry<Long, UserAccount>>
 	public int compare(Map.Entry<Long, UserAccount> o1, Map.Entry<Long, UserAccount> o2) {
 		switch(type) {
 		case COINS:
-			return o1.getValue().getCoins() - o2.getValue().getCoins();
+			return (int)(o1.getValue().getCoins() - o2.getValue().getCoins());
 		case XP:
 			return o1.getValue().getXP() - o2.getValue().getXP();
 		default: return 0;
